Finish Support For Assembly Parsing
Allows multiple statements to be parsed from a single string, separated
by ';'.

All opcodes and registers are read from json files.

A single assembly statement can result in multiple differing possible
binary strings. For instance, 'addRegister' allows using '*' for the
register name to mean any register; this results in multiple but
separate possibilities.

In the event that one operation results in multiple possible binary
statements, and there are multiple assembly statements, the
cross-product of the two vectors is taken and all are searched in the
binary.